backup/first -fina app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Babel==4.0.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
gunicorn==21.2.0
redis==5.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
PyPDF2==3.0.1
